fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
streamlit
transformers
langserve
langchain
langchain_core
langchain_community
langchain-aws
langchain-google-vertexai
sse_starlette
torch
torchvision
torchaudio